Bug fix
Final fix of the bug, diagnosed by @jannamartinek, that causes a runtime error in the cycle model.  This will allow the power output to exceed nameplate capacity, as can be done in realistic operations, but was formerly truncated to capacity by the cycle model.  Capacity is still truncated if a component failure limits the capacity of the plant (Efficiency reductions do not cause a truncation).

Co-Authored-By: jannamartinek <jannamartinek@users.noreply.github.com>