Added calculation of replacement threshold
This update optimizes the replacement threhsold (accoridng to degradation) for each group of mirrors in the solar field.  In the case of linear degradation, the solution takes on a closed form; in the case of accelerated degradation with each passing year (i.e., the acceleration parameter is set to something greater than zero), the optimal replacemnet threhsold is obtained using bisection.

- Added a new parameter (heliostat refurbish cost) to the optical degradation model parameters, since this is now used in the calculation of replacement thresholds instead of just financial outputs.