Adding solarfield optimizer code
Optimizer simulates the failure and repair of heliostat components over the life of the plant, tracking an objective function equal to total revenue loss plus labor and repair costs. The optimzier assumes that the cost function is submodular, and adds staff as long as the return on additional labor investment is positive.